+++ tests/testthat/test-COTAN-getters.R
@@ -1,3 +1,5 @@
+options(COTAN.TorchWarning = NULL)
+
 
 test_that("COTAN getters", {
   raw <- matrix(c(1L,  0L, 4L, 2L, 11L, 0L, 6L, 7L, 0L, 9L,
@@ -11,10 +13,15 @@ test_that("COTAN getters", {
                                sampleCondition = "Test")
   obj <- clean(obj)
   obj <- estimateDispersionBisection(obj)
-  obj <- calculateCoex(obj, actOnCells = FALSE, returnPPFract = TRUE,
-                       optimizeForSpeed = TRUE, deviceStr = "cpu")
-  obj <- calculateCoex(obj, actOnCells = TRUE, returnPPFract = TRUE,
-                       optimizeForSpeed = FALSE)
+
+  suppressWarnings({
+    obj <- calculateCoex(obj, actOnCells = FALSE, returnPPFract = TRUE,
+                         optimizeForSpeed = TRUE, deviceStr = "cpu")
+  })
+  expect_no_warning({
+    obj <- calculateCoex(obj, actOnCells = TRUE, returnPPFract = TRUE,
+                         optimizeForSpeed = FALSE)
+  })
 
   obj <- addClusterization(obj, clName = "Test",
                            clusters = set_names(rep(c(1L, 2L), 10L),