+++ tests/testthat/test-cellsUniformClustering.R
@@ -3,6 +3,7 @@ stopifnot(file.exists(tm))
 
 library(zeallot)
 
+
 test_that("Cell Uniform Clustering", {
   utils::data("test.dataset", package = "COTAN")
 
@@ -42,12 +43,14 @@ test_that("Cell Uniform Clustering", {
   expect_identical(reorderClusterization(obj)[["clusters"]], clusters)
 
   firstCl <- clusters[[1L]]
-  c(isUniform, fracAbove, firstPerc, clSize) %<-%
-    checkClusterUniformity(obj, GDIThreshold = GDIThreshold,
-                           clusterName = paste0("Cluster_", firstCl),
-                           cells = names(clusters)[clusters == firstCl],
-                           optimizeForSpeed = TRUE, deviceStr = "cpu",
-                           saveObj = TRUE, outDir = tm)
+  suppressWarnings({
+    c(isUniform, fracAbove, firstPerc, clSize) %<-%
+      checkClusterUniformity(obj, GDIThreshold = GDIThreshold,
+                             clusterName = paste0("Cluster_", firstCl),
+                             cells = names(clusters)[clusters == firstCl],
+                             optimizeForSpeed = TRUE, deviceStr = "cpu",
+                             saveObj = TRUE, outDir = tm)
+  })
   expect_true(isUniform)
   expect_lte(fracAbove, 0.01)
   expect_lte(firstPerc, GDIThreshold)
@@ -135,7 +138,9 @@ test_that("Cell Uniform Clustering", {
     temp.obj <- dropGenesCells(objCOTAN = obj,
                                cells = getCells(obj)[cellsToDrop])
 
-    temp.obj <- proceedToCoex(temp.obj, cores = 6L, saveObj = FALSE)
+    suppressWarnings({
+      temp.obj <- proceedToCoex(temp.obj, cores = 6L, saveObj = FALSE)
+    })
     gc()
 
     GDI_data <- calculateGDI(temp.obj)