+++ NEWS.md
@@ -4,6 +4,10 @@ Added support for the `torch` library to help with the heavy lifting
 calculations of the *genes' COEX* matrix, with consequent substantial speed-up,
 especially when a **GPU** is available on the system
 
+Allowed user to explicitly **opt-out** from the `torch` library usage:
+COTAN will avoid `torch` commands when the option `"COTAN.UseTorch"` is
+set to `FALSE`
+
 ## COTAN 2.4.1
 
 Fixed bug in `reorderClusterization()` function